+++ lib/src/dual_ring.dart
@@ -3,18 +3,23 @@ import 'dart:math' as math;
 import 'package:flutter/widgets.dart';
 
 class SpinKitDualRing extends StatefulWidget {
-  const SpinKitDualRing({
+  // ignore: prefer_const_constructors_in_immutables
+  SpinKitDualRing({
     Key key,
     @required this.color,
+    this.lineWidth = 7.0,
     this.size = 50.0,
     this.duration = const Duration(milliseconds: 1200),
+    this.controller,
   })  : assert(color != null),
         assert(size != null),
         super(key: key);
 
   final Color color;
+  final double lineWidth;
   final double size;
   final Duration duration;
+  final AnimationController controller;
 
   @override
   _SpinKitDualRingState createState() => _SpinKitDualRingState();
@@ -28,16 +33,14 @@ class _SpinKitDualRingState extends State<SpinKitDualRing>
   @override
   void initState() {
     super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: widget.duration,
-    );
+    _controller = widget.controller ??
+        AnimationController(vsync: this, duration: widget.duration);
     _animation1 = Tween(begin: 0.0, end: 1.0).animate(
       CurvedAnimation(
         parent: _controller,
         curve: const Interval(0.0, 1.0, curve: Curves.linear),
       ),
-    )..addListener(() => setState(() => <String, void>{}));
+    )..addListener(() => setState(() {}));
 
     _controller.repeat();
   }
@@ -60,7 +63,10 @@ class _SpinKitDualRingState extends State<SpinKitDualRing>
           child: SizedBox.fromSize(
             size: Size.square(widget.size),
           ),
-          painter: _DualRingPainter(color: widget.color),
+          painter: _DualRingPainter(
+            paintWidth: widget.lineWidth,
+            color: widget.color,
+          ),
         ),
       ),
     );
@@ -68,31 +74,23 @@ class _SpinKitDualRingState extends State<SpinKitDualRing>
 }
 
 class _DualRingPainter extends CustomPainter {
-  _DualRingPainter({this.weight = 90.0, Color color}) {
-    p.color = color;
-    p.strokeWidth = 10.0;
-    p.style = PaintingStyle.stroke;
-  }
+  _DualRingPainter({
+    this.angle = 90.0,
+    double paintWidth,
+    Color color,
+  }) : ringPaint = Paint()
+          ..color = color
+          ..strokeWidth = paintWidth
+          ..style = PaintingStyle.stroke;
 
-  Paint p = Paint();
-  final double weight;
+  final Paint ringPaint;
+  final double angle;
 
   @override
   void paint(Canvas canvas, Size size) {
-    canvas.drawArc(
-      Rect.fromPoints(Offset.zero, Offset(size.width, size.height)),
-      0.0,
-      getRadian(weight),
-      false,
-      p,
-    );
-    canvas.drawArc(
-      Rect.fromPoints(Offset.zero, Offset(size.width, size.height)),
-      getRadian(180.0),
-      getRadian(weight),
-      false,
-      p,
-    );
+    final rect = Rect.fromPoints(Offset.zero, Offset(size.width, size.height));
+    canvas.drawArc(rect, 0.0, getRadian(angle), false, ringPaint);
+    canvas.drawArc(rect, getRadian(180.0), getRadian(angle), false, ringPaint);
   }
 
   @override