+++ lib/src/fading_cube.dart
@@ -2,12 +2,14 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_spinkit/src/utils.dart';
 
 class SpinKitFadingCube extends StatefulWidget {
+  // ignore: prefer_const_constructors_in_immutables
   SpinKitFadingCube({
     Key key,
     this.color,
     this.size = 50.0,
     this.itemBuilder,
     this.duration = const Duration(milliseconds: 2400),
+    this.controller,
   })  : assert(
             !(itemBuilder is IndexedWidgetBuilder && color is Color) &&
                 !(itemBuilder == null && color == null),
@@ -19,6 +21,7 @@ class SpinKitFadingCube extends StatefulWidget {
   final double size;
   final IndexedWidgetBuilder itemBuilder;
   final Duration duration;
+  final AnimationController controller;
 
   @override
   _SpinKitFadingCubeState createState() => _SpinKitFadingCubeState();
@@ -31,7 +34,8 @@ class _SpinKitFadingCubeState extends State<SpinKitFadingCube>
   @override
   void initState() {
     super.initState();
-    _opacityCtrl = AnimationController(vsync: this, duration: widget.duration)
+    _opacityCtrl = (widget.controller ??
+        AnimationController(vsync: this, duration: widget.duration))
       ..repeat();
   }