+++ lib/src/hour_glass.dart
@@ -8,6 +8,7 @@ class SpinKitHourGlass extends StatefulWidget {
     @required this.color,
     this.size = 50.0,
     this.duration = const Duration(milliseconds: 1200),
+    this.controller,
   })  : assert(color != null),
         assert(size != null),
         super(key: key);
@@ -15,6 +16,7 @@ class SpinKitHourGlass extends StatefulWidget {
   final Color color;
   final double size;
   final Duration duration;
+  final AnimationController controller;
 
   @override
   _SpinKitHourGlassState createState() => _SpinKitHourGlassState();
@@ -28,13 +30,14 @@ class _SpinKitHourGlassState extends State<SpinKitHourGlass>
   @override
   void initState() {
     super.initState();
-    _controller = AnimationController(vsync: this, duration: widget.duration);
+    _controller = widget.controller ??
+        AnimationController(vsync: this, duration: widget.duration);
     _animation1 = Tween(begin: 0.0, end: 8.0).animate(
       CurvedAnimation(
         parent: _controller,
         curve: const Interval(0.0, 1.0, curve: Curves.easeOut),
       ),
-    )..addListener(() => setState(() => <String, void>{}));
+    )..addListener(() => setState(() {}));
 
     _controller.repeat();
   }