+++ lib/src/ripple.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 class SpinKitRipple extends StatefulWidget {
+  // ignore: prefer_const_constructors_in_immutables
   SpinKitRipple({
     Key key,
     this.color,
@@ -8,6 +9,7 @@ class SpinKitRipple extends StatefulWidget {
     this.borderWidth = 6.0,
     this.itemBuilder,
     this.duration = const Duration(milliseconds: 1800),
+    this.controller,
   })  : assert(
             !(itemBuilder is IndexedWidgetBuilder && color is Color) &&
                 !(itemBuilder == null && color == null),
@@ -21,6 +23,7 @@ class SpinKitRipple extends StatefulWidget {
   final double borderWidth;
   final IndexedWidgetBuilder itemBuilder;
   final Duration duration;
+  final AnimationController controller;
 
   @override
   _SpinKitRippleState createState() => _SpinKitRippleState();
@@ -34,7 +37,8 @@ class _SpinKitRippleState extends State<SpinKitRipple>
   @override
   void initState() {
     super.initState();
-    _controller = AnimationController(vsync: this, duration: widget.duration)
+    _controller = (widget.controller ??
+        AnimationController(vsync: this, duration: widget.duration))
       ..repeat();
 
     _animation1 = Tween(begin: 0.0, end: 1.0).animate(
@@ -42,14 +46,14 @@ class _SpinKitRippleState extends State<SpinKitRipple>
         parent: _controller,
         curve: const Interval(0.0, 0.75, curve: Curves.linear),
       ),
-    )..addListener(() => setState(() => <String, void>{}));
+    )..addListener(() => setState(() {}));
 
     _animation2 = Tween(begin: 0.0, end: 1.0).animate(
       CurvedAnimation(
         parent: _controller,
         curve: const Interval(0.25, 1.0, curve: Curves.linear),
       ),
-    )..addListener(() => setState(() => <String, void>{}));
+    )..addListener(() => setState(() {}));
   }
 
   @override