+++ lib/src/three_bounce.dart
@@ -2,12 +2,14 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_spinkit/src/utils.dart';
 
 class SpinKitThreeBounce extends StatefulWidget {
+  // ignore: prefer_const_constructors_in_immutables
   SpinKitThreeBounce({
     Key key,
     this.color,
     this.size = 50.0,
     this.itemBuilder,
     this.duration = const Duration(milliseconds: 1400),
+    this.controller,
   })  : assert(
             !(itemBuilder is IndexedWidgetBuilder && color is Color) &&
                 !(itemBuilder == null && color == null),
@@ -19,6 +21,7 @@ class SpinKitThreeBounce extends StatefulWidget {
   final double size;
   final IndexedWidgetBuilder itemBuilder;
   final Duration duration;
+  final AnimationController controller;
 
   @override
   _SpinKitThreeBounceState createState() => _SpinKitThreeBounceState();
@@ -31,7 +34,8 @@ class _SpinKitThreeBounceState extends State<SpinKitThreeBounce>
   @override
   void initState() {
     super.initState();
-    _scaleCtrl = AnimationController(vsync: this, duration: widget.duration)
+    _scaleCtrl = (widget.controller ??
+        AnimationController(vsync: this, duration: widget.duration))
       ..repeat();
   }